+++ src/components/companies-list/companies-list.js
@@ -9,36 +9,33 @@ import Paper from '@material-ui/core/Paper';
 
 export default function CompaniesListing({ data }) {
   return (
-    <>
-      <h2>Empresas</h2>
-      <TableContainer component={Paper}>
-        <Table aria-label='Companies list'>
-          <TableHead>
-            <TableRow>
-              <TableCell>Nome</TableCell>
-              <TableCell align='right'>Telefone</TableCell>
-              <TableCell align='right'>Desconto&nbsp;</TableCell>
-              <TableCell align='right'>CNPJ</TableCell>
-              <TableCell align='right'>Ativo</TableCell>
+    <TableContainer component={Paper}>
+      <Table aria-label='Companies list'>
+        <TableHead>
+          <TableRow>
+            <TableCell>Nome</TableCell>
+            <TableCell align='right'>Telefone</TableCell>
+            <TableCell align='right'>Desconto&nbsp;</TableCell>
+            <TableCell align='right'>CNPJ</TableCell>
+            <TableCell align='right'>Ativo</TableCell>
+          </TableRow>
+        </TableHead>
+        <TableBody>
+          {data.map((company) => (
+            <TableRow key={company.name}>
+              <TableCell component='th' scope='row'>
+                {company.name}
+              </TableCell>
+              <TableCell align='right'>{company.phone}</TableCell>
+              <TableCell align='right'>{company.discount}</TableCell>
+              <TableCell align='right'>{company.cnpj}</TableCell>
+              <TableCell align='right'>
+                {company.active ? 'Sim' : 'Não'}
+              </TableCell>
             </TableRow>
-          </TableHead>
-          <TableBody>
-            {data.map((company) => (
-              <TableRow key={company.name}>
-                <TableCell component='th' scope='row'>
-                  {company.name}
-                </TableCell>
-                <TableCell align='right'>{company.phone}</TableCell>
-                <TableCell align='right'>{company.discount}</TableCell>
-                <TableCell align='right'>{company.cnpj}</TableCell>
-                <TableCell align='right'>
-                  {company.active ? 'Sim' : 'Não'}
-                </TableCell>
-              </TableRow>
-            ))}
-          </TableBody>
-        </Table>
-      </TableContainer>
-    </>
+          ))}
+        </TableBody>
+      </Table>
+    </TableContainer>
   );
 }
