+++ src/components/form_error/styles.js
@@ -1,0 +1,11 @@
+import styled from 'styled-components';
+import theme from '../../styles/theme';
+
+export const Container = styled.div`
+  background-color: ${theme.palette.warning.light};
+  color: ${theme.palette.warning.dark};
+  width: 100%;
+  padding: 10px 20px;
+  border-radius: 4px;
+  margin-bottom: 30px;
+`;
