+++ src/pages/companies-list/companies-list.js
@@ -3,7 +3,7 @@ import CompaniesListing from '../../components/companies-list/companies-list';
 import Template from '../../components/template/template.component';
 import { getCompanies } from '../../service/company';
 import { toast } from 'react-toastify';
-import { CircularProgress, Box } from '@material-ui/core';
+import { CircularProgress, Box, Button } from '@material-ui/core';
 
 const CompaniesList = () => {
   const [companies, setCompanies] = useState([]);
@@ -40,7 +40,20 @@ const CompaniesList = () => {
         </Box>
       );
 
-    return <CompaniesListing data={companies} />;
+    return (
+      <>
+        <h2>Empresas</h2>
+        <Button
+          variant='contained'
+          color='primary'
+          href='/companies/new'
+          style={{ marginBottom: '20px' }}
+        >
+          Cadastrar Empresa
+        </Button>
+        <CompaniesListing data={companies} />
+      </>
+    );
   };
 
   return <Template>{render()}</Template>;
