+++ src/pages/companies/company_register.component.js
@@ -1,6 +1,6 @@
 import React, { useState } from 'react';
 import Template from '../../components/template/template.component';
-import CompanyForm from './company_form.component';
+import CompanyForm from './form/company_form.component';
 import FormErrors from '../../components/form_error/form_errors.component';
 import { registerCompany } from '../../service/company';
 import { useHistory } from 'react-router-dom';
@@ -27,7 +27,7 @@ const CompanyRegister = () => {
 
   return (
     <Template>
-      <h1>Cadastrar empresa</h1>
+      <h2>Cadastrar empresa</h2>
       {!!formErrors.length && !loading && (
         <FormErrors action='registro' errors={formErrors} />
       )}
