+++ src/pages/companies/form/styles.js
@@ -1,0 +1,8 @@
+import styled from 'styled-components';
+
+export const Container = styled.div`
+  background: #fff;
+  padding: 0 30px;
+  padding-bottom: 30px;
+  border-radius: 5px;
+`;
